requirements.txt
flask==2.2.5
pandas==2.3.3
numpy==2.3.4
requests==2.32.5
python-dotenv==1.2.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.35
GeoAlchemy2==0.15.2
apache-airflow==2.10.2
scikit-learn==1.5.2